server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
paho-mqtt==2.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.12.1
pytest==7.4.3
